# AnotherMe/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnotherMe Gateway API",
    "version": "0.3.0",
    "description": "Unified backend gateway for AnotherMe Web and Mobile clients.\n\nCapabilities follow the registry defined in `agents/foundation/capability_registry.py`.\nValid capability ids: `course_generate`, `problem_video_generate`, `quiz_practice`,\n`interactive_demo`, `ai_tutor_chat`, `deep_solve`, `deep_research`,\n`math_animator`, `visualize`, `co_writer`.\n\nP1 changes:\n  - Capability endpoints (P0-3) are now wired.\n  - Global error envelope (`{ error_code, message, request_id, details? }`).\n  - Live-book / co-writer / courses routes are first-class.\n\nAuth: Phase 0 uses static `Bearer <GATEWAY_API_TOKEN>`. Phase 2 will switch to Supabase JWT.",
    "contact": {
      "name": "AnotherMe",
      "url": "https://github.com/anotherme/anotherme"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Phase 0: static token from GATEWAY_API_TOKEN env var. Phase 2: Supabase JWT."
      }
    },
    "schemas": {
...
      "ImageGenerationRequest": {
        "type": "object",
        "required": ["prompt"],
        "properties": {
          "prompt": { "type": "string" },
          "negative_prompt": { "type": "string" },
          "width": { "type": "integer", "default": 1024 },
          "height": { "type": "integer", "default": 1024 },
          "aspect_ratio": { "type": "string" },
          "style": { "type": "string" }
        }
      },
      "ImageGenerationResult": {
        "type": "object",
        "required": ["base64", "format", "width", "height"],
        "properties": {
          "base64": { "type": "string" },
          "format": { "type": "string" },
          "width": { "type": "integer" },
          "height": { "type": "integer" }
        }
      },
      "ImageGenerationResponse": {
        "type": "object",
        "required": ["success", "result"],
        "properties": {
          "success": { "type": "boolean" },
          "result": { "$ref": "#/components/schemas/ImageGenerationResult" }
        }
      },
      "TTSRequest": {
        "type": "object",
        "required": ["text", "audio_id", "tts_voice"],
        "properties": {
          "text": { "type": "string" },
          "audio_id": { "type": "string" },
          "tts_provider_id": { "type": "string", "default": "openai-tts" },
          "tts_voice": { "type": "string", "default": "alloy" },
          "tts_model_id": { "type": "string" },
          "tts_speed": { "type": "number", "default": 1.0 },
          "tts_api_key": { "type": "string" },
          "tts_base_url": { "type": "string" }
        }
      },
      "TTSResponse": {
        "type": "object",
        "required": ["success", "url"],
        "properties": {
          "success": { "type": "boolean" },
          "url": { "type": "string" },
          "audio_id": { "type": "string" }
        }
      },
      "VideoGenerationRequest": {
        "type": "object",
        "required": ["prompt"],
        "properties": {
          "prompt": { "type": "string" },
          "duration": { "type": "integer", "default": 5 },
          "aspect_ratio": { "type": "string", "default": "16:9" },
          "resolution": { "type": "string", "default": "720p" }
        }
      },
      "VideoGenerationResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean" },
          "url": { "type": "string" }
        }
      },
...
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/generate/image": {
      "post": {
        "tags": ["media"],
        "summary": "Generate an image (Seedream/Qwen/NanoBanana/etc.)",
        "parameters": [
          { "name": "x-image-provider", "in": "header", "required": false, "schema": { "type": "string", "default": "seedream" } },
          { "name": "x-api-key", "in": "header", "required": false, "schema": { "type": "string" } },
          { "name": "x-base-url", "in": "header", "required": false, "schema": { "type": "string" } },
          { "name": "x-image-model", "in": "header", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImageGenerationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image generated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ImageGenerationResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/generate/tts": {
      "post": {
        "tags": ["media"],
        "summary": "Generate TTS audio (OpenAI/Azure/MiniMax/etc.)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TTSRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio generated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TTSResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/generate/video": {
      "post": {
        "tags": ["media"],
        "summary": "Generate a video (Seedance/Kling/Veo/MiniMax/etc.)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VideoGenerationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video generated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VideoGenerationResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}